+++ src/lua/CommunityBalanceMod/Scripts/GUIBetaBalanceChangelogData.lua
@@ -46,10 +46,10 @@ on the official discord to let me and the team know what you think! Below are th
   - Shells/Veils/Spurs are able to selfheal/selfcloak or move (buff)
   - Drifter hallucination ability replaced with Cloaking Haze ability (server performance)
 
-# Changes between Revision 2.0.3 and 2.0.2: (2023/12/24)
+# Changes between Revision 2.0.3 and 2.0.2: (2023/12/??)
 ## ALIEN
 ### Hives
-  - Reduced the additional 10% HP buff for maturity hives to 5%
+  - Reduced the additional 10% HP buff for full maturity hives to 5%
 
 ### Stab
   - Stab research cost reduced from 25 to 20 tres
@@ -58,7 +58,7 @@ on the official discord to let me and the team know what you think! Below are th
 ### Fortress Shade
   - Hallucinations time duration removed
   - Able to transform into a random different structure
-  - Added HP bars on hallucinations
+  - Added HP bars to hallucinations
 
 ### Whips
   - Increased turning speed before moving
@@ -66,13 +66,14 @@ on the official discord to let me and the team know what you think! Below are th
   - Reduced the cost of bile splash from 3 to 2 tres
 
 ### Fortress Structures
-  - Converted 66% of their armor into HP (huge nerf to healing)
+  - Converted 50% of their armor into HP (nerf to healing)
   - Increased cooldown for all fortress abilities from 10 to 15 seconds
-  - reduced the cost of all fortress abilities from 3 to 2 tres
+  - Reduced the cost of all fortress abilities from 3 to 2 tres
   - Fixed a bug which allowed abilities to being cast while under fire
   - No longer take 3x the damage off infestation in comparison to normal structures
 
 ### Drifter
+  - Increased mucous area of effect to the same size as enzymes
   - Doesnt follow echoed unfinished structures over the entire map anymore
   
 ## MARINE